Fix whofaved_deviation by changing to get_data
`whofaved_deviation` will always return with the default pagination parameters -- `offset = 0`, `limit = 10` -- as written, because the parameters to the request are written to the `post_data` struct even though this is a `GET` request. 

That was probably a small oversight that should be fixed with this one line of change.